Refactor route data
