+++ src/main/java/arcs/commands/route/AddRouteCommand.java
@@ -2,8 +2,7 @@ package arcs.commands.route;
 
 import arcs.commands.Command;
 import arcs.commands.CommandResult;
-import arcs.data.Route;
-import arcs.data.exception.ArcsException;
+import arcs.data.Route.Route;
 import arcs.data.exception.DuplicateDataException;
 
 import java.util.ArrayList;
