+++ src/main/java/arcs/commands/route/ListRouteCommand.java
@@ -2,7 +2,7 @@ package arcs.commands.route;
 
 import arcs.commands.Command;
 import arcs.commands.CommandResult;
-import arcs.data.Route;
+import arcs.data.Route.Route;
 
 import java.util.ArrayList;
