+++ src/test/java/arcs/data/RouteManagerTest.java
@@ -1,5 +1,7 @@
 package arcs.data;
 
+import arcs.data.Route.Route;
+import arcs.data.Route.RouteManager;
 import arcs.data.exception.DuplicateDataException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
